requirements.txt
# requirements.txt
openai==1.12.0
langchain==0.1.9
langchain-openai==0.0.8
chromadb==0.4.24
faiss-cpu==1.7.4
streamlit==1.31.0
fastapi==0.109.2
uvicorn==0.27.1
pypdf2==3.0.1
python-docx==1.1.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
plotly==5.18.0
python-multipart==0.0.9
aiofiles==23.2.0